refactored the structure of the test folder, only have one package now.
